This project was upgraded with external_updater.
Usage: tools/external_updater/updater.sh update <absolute path to project>
For more info, check https://cs.android.com/android/platform/superproject/+/main:tools/external_updater/README.md

Test: TreeHugger

+++ release-process/rr-testing.sh
@@ -1,10 +1,24 @@
 # Bash script to build rr and run tests.
 #
-# Requires variables to be set:
+# Requires variables and functions to be set. See test-system.py.
 # $git_revision : git revision to check out, build and test
+# $staticlibs : TRUE or FALSE to build with static libs
 # $build_dist : 1 if we should build dist packages, 0 otherwise
+# $test_firefox : 1 to run firefox tests, 0 to skip
+# $ctest_options : options to pass to ctest, e.g to exclude certain tests
+# setup_commands : function to setup environment, e.g. 'apt update'
+# install_build_deps : function to install dependencies required to build rr
+# install_test_deps : function to install dependencies required by tests
+
+set -x # echo commands
+set -e # default to exiting on error"
 
 uname -a
+
+setup_commands
+install_build_deps
+
+install_test_deps & # job %1
 
 # Free up space before we (re)start
 
@@ -16,12 +30,12 @@ git checkout $git_revision
 rm -rf ~/obj || true
 mkdir ~/obj
 cd ~/obj
-cmake -G Ninja -DCMAKE_BUILD_TYPE=RELEASE -Dstaticlibs=TRUE -Dstrip=TRUE ../rr
+cmake -G Ninja -DCMAKE_BUILD_TYPE=RELEASE -Dstaticlibs=$staticlibs -Dstrip=TRUE ../rr
 ninja
 
 # Test deps are installed in parallel with our build.
 # Make sure that install has finished before running tests
-wait_for_test_deps
+wait %1
 
 # Enable perf events for rr
 echo 0 | sudo tee /proc/sys/kernel/perf_event_paranoid
@@ -69,12 +83,17 @@ function xvnc-runner { CMD=$1 EXPECT=$2
   echo PASSED: $CMD
 }
 
-rm -rf /tmp/firefox-profile || true
-mkdir /tmp/firefox-profile
-xvnc-runner "firefox --profile /tmp/firefox-profile $HOME/rr/release-process/test-data/test.html" "rr Test Page"
+if [[ $test_firefox == 1 ]]; then
+  rm -rf /tmp/firefox /tmp/firefox-profile || true
+  mkdir /tmp/firefox-profile
+  ( cd /tmp; curl -L 'https://download.mozilla.org/?product=firefox-latest&os=linux64&lang=en-US' | tar -jxf - )
+  xvnc-runner "/tmp/firefox/firefox --profile /tmp/firefox-profile $HOME/rr/release-process/test-data/test.html" "rr Test Page"
+fi
 
-rm -rf ~/.config/libreoffice || true
-xvnc-runner "libreoffice $HOME/rr/release-process/test-data/rr-test-doc.odt" "rr-test-doc.odt"
+if [[ $test_libreoffice == 1 ]]; then
+  rm -rf ~/.config/libreoffice || true
+  xvnc-runner "libreoffice $HOME/rr/release-process/test-data/rr-test-doc.odt" "rr-test-doc.odt"
+fi
 
 if [[ $build_dist != 0 ]]; then
   make -j`nproc` dist
